+++ sdk/azurestackhci/arm-azurestackhci/src/azureStackHCIClient.ts
@@ -16,7 +16,6 @@ import {
 import * as coreAuth from "@azure/core-auth";
 import {
   GalleryImagesOperationsImpl,
-  OperationsImpl,
   LogicalNetworksOperationsImpl,
   MarketplaceGalleryImagesOperationsImpl,
   NetworkInterfacesOperationsImpl,
@@ -29,7 +28,6 @@ import {
 } from "./operations";
 import {
   GalleryImagesOperations,
-  Operations,
   LogicalNetworksOperations,
   MarketplaceGalleryImagesOperations,
   NetworkInterfacesOperations,
@@ -88,7 +86,7 @@ export class AzureStackHCIClient extends coreClient.ServiceClient {
       credential: credentials
     };
 
-    const packageDetails = `azsdk-js-arm-azurestackhci/4.0.0-beta.1`;
+    const packageDetails = `azsdk-js-arm-azurestackhci/4.0.0`;
     const userAgentPrefix =
       options.userAgentOptions && options.userAgentOptions.userAgentPrefix
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`
@@ -141,9 +139,8 @@ export class AzureStackHCIClient extends coreClient.ServiceClient {
 
     // Assigning values to Constant parameters
     this.$host = options.$host || "https://management.azure.com";
-    this.apiVersion = options.apiVersion || "2023-09-01-preview";
+    this.apiVersion = options.apiVersion || "2024-01-01";
     this.galleryImagesOperations = new GalleryImagesOperationsImpl(this);
-    this.operations = new OperationsImpl(this);
     this.logicalNetworksOperations = new LogicalNetworksOperationsImpl(this);
     this.marketplaceGalleryImagesOperations = new MarketplaceGalleryImagesOperationsImpl(
       this
@@ -193,7 +190,6 @@ export class AzureStackHCIClient extends coreClient.ServiceClient {
   }
 
   galleryImagesOperations: GalleryImagesOperations;
-  operations: Operations;
   logicalNetworksOperations: LogicalNetworksOperations;
   marketplaceGalleryImagesOperations: MarketplaceGalleryImagesOperations;
   networkInterfacesOperations: NetworkInterfacesOperations;